refactors test artifacts
